chore(frontend): snapshot minimul Streamlit app as v0 and scaffolding v1 file
